fix(#179): review fixes — extract CommunityLookup, URL-encode slugs, tighten tests
- Extract duplicated buildCommunityLookup() into src/Support/CommunityLookup.php
- URL-encode community_slug in card template hrefs for XSS defense
- Replace always-true Playwright assertions (>= 0) with meaningful checks
