+++ src/Controller/GroupController.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Minoo\Controller;
 
+use Minoo\Support\CommunityLookup;
 use Symfony\Component\HttpFoundation\Request as HttpRequest;
 use Twig\Environment;
 use Waaseyaa\Access\AccountInterface;
@@ -38,7 +39,7 @@ final class GroupController
         });
         $groups = array_values($groups);
 
-        $communities = $this->buildCommunityLookup($groups);
+        $communities = CommunityLookup::build($this->entityTypeManager, $groups);
 
         $html = $this->twig->render('groups.html.twig', [
             'path' => '/groups',
@@ -80,30 +81,5 @@ final class GroupController
             content: $html,
             statusCode: $group !== null ? 200 : 404,
         );
-    }
-
-    /** @param list<\Waaseyaa\Entity\EntityInterface> $entities */
-    private function buildCommunityLookup(array $entities): array
-    {
-        $communityIds = array_filter(array_unique(array_map(
-            fn ($e) => $e->get('community_id'),
-            $entities
-        )));
-
-        if ($communityIds === []) {
-            return [];
-        }
-
-        $communityStorage = $this->entityTypeManager->getStorage('community');
-        $communities = $communityStorage->loadMultiple($communityIds);
-        $lookup = [];
-        foreach ($communities as $community) {
-            $lookup[(string) $community->id()] = [
-                'name' => $community->get('name') ?? $community->label(),
-                'slug' => $community->get('slug'),
-            ];
-        }
-
-        return $lookup;
     }
 }
