+++ src/Controller/TeachingController.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Minoo\Controller;
 
+use Minoo\Support\CommunityLookup;
 use Symfony\Component\HttpFoundation\Request as HttpRequest;
 use Twig\Environment;
 use Waaseyaa\Access\AccountInterface;
@@ -39,7 +40,7 @@ final class TeachingController
         });
         $teachings = array_values($teachings);
 
-        $communities = $this->buildCommunityLookup($teachings);
+        $communities = CommunityLookup::build($this->entityTypeManager, $teachings);
 
         $html = $this->twig->render('teachings.html.twig', [
             'path' => '/teachings',
@@ -82,30 +83,5 @@ final class TeachingController
             content: $html,
             statusCode: $teaching !== null ? 200 : 404,
         );
-    }
-
-    /** @param list<\Waaseyaa\Entity\EntityInterface> $entities */
-    private function buildCommunityLookup(array $entities): array
-    {
-        $communityIds = array_filter(array_unique(array_map(
-            fn ($e) => $e->get('community_id'),
-            $entities
-        )));
-
-        if ($communityIds === []) {
-            return [];
-        }
-
-        $communityStorage = $this->entityTypeManager->getStorage('community');
-        $communities = $communityStorage->loadMultiple($communityIds);
-        $lookup = [];
-        foreach ($communities as $community) {
-            $lookup[(string) $community->id()] = [
-                'name' => $community->get('name') ?? $community->label(),
-                'slug' => $community->get('slug'),
-            ];
-        }
-
-        return $lookup;
     }
 }
