+++ templates/components/teaching-card.html.twig
@@ -5,7 +5,7 @@
     <p class="card__meta">{{ author }}</p>
   {% endif %}
   {% if community_name is defined and community_name %}
-    <p class="card__meta card__community"><a href="/communities/{{ community_slug }}">{{ community_name }}</a></p>
+    <p class="card__meta card__community"><a href="/communities/{{ community_slug|url_encode }}">{{ community_name }}</a></p>
   {% endif %}
   {% if excerpt is defined and excerpt %}
     <p class="card__body">{{ excerpt }}</p>
