+++ tests/playwright/community-contact.spec.ts
@@ -1,49 +1,40 @@
 import { test, expect } from '@playwright/test';
 
 test.describe('Community contact card', () => {
-  test('community detail page renders leadership section', async ({ page }) => {
-    // Navigate to the listing first to find a community with data
+  test('community detail page renders about section', async ({ page }) => {
     await page.goto('/communities');
     const cards = page.locator('a.atlas-card');
     const count = await cards.count();
     test.skip(count === 0, 'No community cards available — requires seeded data');
 
-    // Click the first card to get a real community slug
     await cards.first().click();
     await expect(page.locator('h1')).toBeVisible();
     expect(page.url()).toMatch(/\/communities\/.+/);
 
-    // If leadership section labels exist, verify at least one is present
-    // This is a soft check — NC contact data may not be available in CI
+    // Every community detail page must have at least the "About" section label
     const sectionLabels = page.locator('.atlas-section__label');
     const labelCount = await sectionLabels.count();
-    expect(labelCount).toBeGreaterThanOrEqual(0);
+    expect(labelCount).toBeGreaterThan(0);
   });
 
-  test('community detail page renders without NC data gracefully', async ({ page }) => {
-    // Navigate to the listing to find any community
+  test('community detail page degrades gracefully without NC data', async ({ page }) => {
     await page.goto('/communities');
     const cards = page.locator('a.atlas-card');
     const count = await cards.count();
     test.skip(count === 0, 'No community cards available — requires seeded data');
 
-    // Visit a community page — it should always render (NC data optional)
     await cards.first().click();
     await expect(page.locator('h1')).toBeVisible();
 
-    // The page should not show leader cards when NC data is absent
-    // atlas-leader-card is only rendered when NC returns contact data
-    const leaderCards = page.locator('.atlas-leader-card');
-    const leaderCount = await leaderCards.count();
-    // Count may be 0 (no NC data) or more (NC data present) — both are valid
-    expect(leaderCount).toBeGreaterThanOrEqual(0);
+    // Page renders successfully regardless of NC data availability
+    // The hero section and about section should always be present
+    await expect(page.locator('.atlas-detail-hero__name')).toBeVisible();
   });
 
   test('community contact card section is absent for 404 communities', async ({ page }) => {
     const response = await page.goto('/communities/nonexistent-community-xyz');
     expect(response?.status()).toBe(404);
 
-    // No contact card should appear on error pages
     const leaderCards = page.locator('.atlas-leader-card');
     await expect(leaderCards).toHaveCount(0);
   });
